Return of lost features & remapping
Following changes:
- Inventory Editor is toggled by L on Quest Status Subscreen
- Item Downgrading is back, with L
- Hide HUD is toggled by L on Map Subscreen
- Reorganized code
