+++ ASM/c/buttons.c
@@ -5,11 +5,67 @@ typedef void(*playsfx_t)(uint16_t sfx, z64_xyzf_t *unk_00_, int8_t unk_01_ , flo
 extern uint8_t CFG_HUD_LAYOUT;
 extern uint8_t CFG_HIDE_HUD_ENABLED;
 extern uint8_t CFG_B_BUTTON_ITEM_ENABLED;
+extern uint8_t CFG_DPAD_ENABLED;
+extern uint8_t CFG_FPS_ENABLED;
+extern uint8_t CFG_INVENTORY_EDITOR_ENABLED;
 
 uint8_t HUD_HIDE		= 0;
 uint8_t HUD_HEARTS_HIDE	= 1;
 uint8_t HUD_COUNTER		= 0;
+uint8_t BLOCK			= 0;
+uint8_t PRESSED_R		= 0;
+uint8_t PRESSED_Z		= 0;
 uint16_t LAST_SCENE		= 0xFFFF;
+
+void handle_buttons() {
+	pad_t pad_pressed = z64_game.common.input[0].pad_pressed;
+	
+	handle_layout();
+	handle_hud();
+	set_b_button(pad_pressed);
+	
+	if (z64_game.pause_ctxt.state == 6 && pad_pressed.s && z64_game.pause_ctxt.unk_02_[1])
+		z64_game.pause_ctxt.unk_02_[1] = 0;
+}
+
+
+void handle_l_button() {
+	if (z64_game.pause_ctxt.state != 0 && z64_game.pause_ctxt.state != 6)
+		return;
+	
+	pad_t pad_released = z64_game.common.input[0].pad_released;
+	
+	if (CFG_DPAD_ENABLED == 2) {
+		if (z64_game.common.input[0].raw.pad.r)
+			PRESSED_R = 1;
+		if (CFG_FPS_ENABLED && z64_game.common.input[0].raw.pad.z)
+			PRESSED_Z = 1;
+		if (pad_released.l && !PRESSED_R && !PRESSED_Z) {
+			toggle_minimap();
+			hide_hud();
+			inventory_editor();
+			handle_downgrading();
+		}
+		if (!z64_game.common.input[0].raw.pad.l)
+			PRESSED_R = PRESSED_Z = 0;
+	}
+	
+	if (z64_game.common.input[0].pad_pressed.l && !z64_game.common.input[0].raw.pad.r && CFG_DPAD_ENABLED == 2)
+		BLOCK = 1;
+	else if (z64_game.common.input[0].pad_pressed.l && !z64_game.common.input[0].raw.pad.z && (CFG_FPS_ENABLED || CFG_INVENTORY_EDITOR_ENABLED) )
+		BLOCK = 1;
+	if (!z64_game.common.input[0].raw.pad.l)
+		BLOCK = 0;
+	
+	if (BLOCK) {
+		if (CFG_DPAD_ENABLED == 2)
+			z64_game.common.input[0].raw.pad.r = z64_game.common.input[0].pad_pressed.r = 0;
+		if (CFG_FPS_ENABLED || CFG_INVENTORY_EDITOR_ENABLED)
+			z64_game.common.input[0].raw.pad.z = z64_game.common.input[0].pad_pressed.z = 0;
+	}
+	
+	z64_game.common.input[0].pad_pressed.l = 0;	
+}
 
 void handle_layout() {
 	if (CFG_HUD_LAYOUT == 0 || !CAN_DRAW_HUD || z64_game.scene_index == LAST_SCENE)
@@ -114,6 +170,22 @@ void handle_layout() {
 	z64_gameinfo.c_up_icon_y		+= c_up_y;
 }
 
+void toggle_minimap() {
+	if (z64_game.pause_ctxt.state != 0 || HUD_HIDE)
+		return;
+	if (z64_dungeon_scene != 0xFFFF) {
+		if (!z64_file.dungeon_items[z64_dungeon_scene].map)
+			return;
+	}		
+	else if (z64_has_minimap == 0xFFFF)
+		return;
+	
+	z64_gameinfo.minimap_disabled ^= 1;
+	if (z64_gameinfo.minimap_disabled)
+		z64_playsfx(0x4813, (z64_xyzf_t*)0x80104394, 0x04, (float*)0x801043A0, (float*)0x801043A0, (float*)0x801043A8);
+	else z64_playsfx(0x4814, (z64_xyzf_t*)0x80104394, 0x04, (float*)0x801043A0, (float*)0x801043A0, (float*)0x801043A8);
+}
+
 void handle_hud() {
 	if (HUD_HIDE && (z64_game.pause_ctxt.state == 0 || z64_game.pause_ctxt.state == 0x1A || z64_game.pause_ctxt.state == 0x1B) ) {
 		if (z64_game.hud_alpha_channels.minimap != 0) {
@@ -142,44 +214,66 @@ void handle_hud() {
 	}
 }
 
-void handle_buttons() {
-	if (z64_game.pause_ctxt.unk_02_[1] != 0)
+void hide_hud() {
+	if (!CFG_HIDE_HUD_ENABLED || z64_game.pause_ctxt.unk_02_[1] != 0)
 		return;
-	pad_t pad_pressed = z64_game.common.input[0].pad_pressed;
 	
-	if (pad_pressed.a && z64_game.pause_ctxt.screen_idx == 0 && z64_game.pause_ctxt.cursor_pos != 0x0A && z64_game.pause_ctxt.cursor_pos != 0x0B) {
-		if (z64_game.pause_ctxt.item_cursor >= Z64_SLOT_STICK && z64_game.pause_ctxt.item_cursor <= Z64_SLOT_CHILD_TRADE && CFG_B_BUTTON_ITEM_ENABLED) {
-			z64_item_t item = 0xFF;
+	if (z64_game.pause_ctxt.state == 6 && z64_game.pause_ctxt.screen_idx == 1) {
+		HUD_HIDE ^= 1;
+		if (HUD_HIDE)
+			z64_playsfx(0x4813, (z64_xyzf_t*)0x80104394, 0x04, (float*)0x801043A0, (float*)0x801043A0, (float*)0x801043A8);
+		else z64_playsfx(0x4814, (z64_xyzf_t*)0x80104394, 0x04, (float*)0x801043A0, (float*)0x801043A0, (float*)0x801043A8);
+	}
+	else if (z64_game.pause_ctxt.state == 0 && HUD_HIDE)
+		HUD_HEARTS_HIDE ^= 1;
+}
+
+void set_b_button(pad_t pad_pressed) {
+	if (!CFG_B_BUTTON_ITEM_ENABLED || !pad_pressed.a || z64_game.pause_ctxt.screen_idx != 0 || z64_game.pause_ctxt.unk_02_[1] != 0 || z64_game.pause_ctxt.cursor_pos == 0x0A || z64_game.pause_ctxt.cursor_pos == 0x0B || z64_game.pause_ctxt.item_cursor <= Z64_SLOT_STICK || z64_game.pause_ctxt.item_cursor >= Z64_SLOT_CHILD_TRADE)
+		return;
+	
+	z64_item_t item = 0xFF;
+	for (uint8_t i=0; i<18; i++)
+		if (z64_game.pause_ctxt.item_cursor == i)
+			item = z64_file.items[i];
 			
-			for (uint8_t i=0; i<18; i++)
-				if (z64_game.pause_ctxt.item_cursor == i)
-					item = z64_file.items[i];
-			
-			if (!z64_file.link_age) {
-				if		(item == Z64_ITEM_STICK)		item = 0xFF;
-				else if (item == Z64_ITEM_SLINGSHOT)	item = 0xFF;
-				else if (item == Z64_ITEM_BOOMERANG)	item = 0xFF;
-				else if (item == Z64_ITEM_BEANS)		item = 0xFF;
-			}
-			else {
-				if		(item == Z64_ITEM_BOW)			item = 0xFF;
-				else if (item == Z64_ITEM_FIRE_ARROW)	item = 0xFF;
-				else if (item == Z64_ITEM_ICE_ARROW)	item = 0xFF;
-				else if (item == Z64_ITEM_LIGHT_ARROW)	item = 0xFF;
-				else if (item == Z64_ITEM_HOOKSHOT)		item = 0xFF;
-				else if (item == Z64_ITEM_LONGSHOT)		item = 0xFF;
-				else if (item == Z64_ITEM_HAMMER)		item = 0xFF;
-			}
-			
-			if (item != Z64_ITEM_SLINGSHOT && item != Z64_ITEM_BOW && item != Z64_ITEM_BOMBCHU && item != Z64_ITEM_LENS && item <= Z64_ITEM_NAYRUS_LOVE) {
-				z64_file.button_items[0] = item;
-				if (!z64_file.link_age)
-					z64_file.adult_button_items[0] = item;
-				else if	(z64_file.link_age)
-					z64_file.child_button_items[0] = item;
-				z64_UpdateItemButton(&z64_game, 0);
-				z64_playsfx(0x4808, (z64_xyzf_t*)0x80104394, 0x04, (float*)0x801043A0, (float*)0x801043A0, (float*)0x801043A8);
-			}
-		}
+	if (!z64_file.link_age) {
+		if		(item == Z64_ITEM_STICK)		item = 0xFF;
+		else if (item == Z64_ITEM_SLINGSHOT)	item = 0xFF;
+		else if (item == Z64_ITEM_BOOMERANG)	item = 0xFF;
+		else if (item == Z64_ITEM_BEANS)		item = 0xFF;
+	}
+	else {
+		if		(item == Z64_ITEM_BOW)			item = 0xFF;
+		else if (item == Z64_ITEM_FIRE_ARROW)	item = 0xFF;
+		else if (item == Z64_ITEM_ICE_ARROW)	item = 0xFF;
+		else if (item == Z64_ITEM_LIGHT_ARROW)	item = 0xFF;
+		else if (item == Z64_ITEM_HOOKSHOT)		item = 0xFF;
+		else if (item == Z64_ITEM_LONGSHOT)		item = 0xFF;
+		else if (item == Z64_ITEM_HAMMER)		item = 0xFF;
+	}
+	
+	if (item != Z64_ITEM_SLINGSHOT && item != Z64_ITEM_BOW && item != Z64_ITEM_BOMBCHU && item != Z64_ITEM_LENS && item <= Z64_ITEM_NAYRUS_LOVE) {
+		z64_file.button_items[0] = item;
+		if (!z64_file.link_age)
+			z64_file.adult_button_items[0] = item;
+		else if	(z64_file.link_age)
+			z64_file.child_button_items[0] = item;
+		z64_UpdateItemButton(&z64_game, 0);
+		z64_playsfx(0x4808, (z64_xyzf_t*)0x80104394, 0x04, (float*)0x801043A0, (float*)0x801043A0, (float*)0x801043A8);
+	}
+}
+
+void inventory_editor() {
+	if (!CFG_INVENTORY_EDITOR_ENABLED || z64_game.pause_ctxt.state != 6)
+		return;
+	
+	if (z64_game.pause_ctxt.unk_02_[1] == 0 && z64_game.pause_ctxt.screen_idx == 2) {
+		z64_game.pause_ctxt.unk_02_[1] = 2;
+		z64_playsfx(0x4813, (z64_xyzf_t*)0x80104394, 0x04, (float*)0x801043A0, (float*)0x801043A0, (float*)0x801043A8);
+	}
+	else if (z64_game.pause_ctxt.unk_02_[1] == 2) {
+		z64_game.pause_ctxt.unk_02_[1] = 0;
+		z64_playsfx(0x4814, (z64_xyzf_t*)0x80104394, 0x04, (float*)0x801043A0, (float*)0x801043A0, (float*)0x801043A8);
 	}
 }
