+++ ASM/c/z64_extended.h
@@ -3,6 +3,18 @@
 
 #include "z64.h"
 
+#define BLOCK_DPAD (0x00000001 | \
+	                0x00000002 | \
+                    0x00000080 | \
+                    0x00000400 | \
+                    0x10000000 | \
+                    0x20000000)
+
+#define CAN_USE_DPAD        (((z64_link.state_flags_1 & BLOCK_DPAD) == 0) && \
+                            ((uint32_t)z64_ctxt.state_dtor==z64_state_ovl_tab[3].vram_dtor) && \
+                            (z64_file.game_mode == 0) && \
+                            ((z64_event_state_1 & 0x20) == 0))
+
 typedef enum {
 	DPAD_NULL = -1,
 	DPAD_SWORD = 1,
@@ -39,7 +51,9 @@ typedef enum {
 	DPAD_CHILD_SET2_RIGHT,
 	DPAD_CHILD_SET2_DOWN,
 	DPAD_CHILD_SET2_LEFT,
-	DOWNGRADING,
+	DOWNGRADE_OCARINA,
+	DOWNGRADE_HOOKSHOT,
+	DOWNGRADE_GIANTS_KNIFE,
 } unused_sram_t;
 
 typedef enum {
